Reword the helpers test because I keep copying it

+++ test/cli/autocorrect-helpers/autocorrect-helpers.sh
@@ -1,3 +1,5 @@
+#!/usr/bin/env bash
+
 tmp="$(mktemp -d)"
 infile="test/cli/autocorrect-helpers/autocorrect-helpers.rb"
 cp "$infile" "$tmp"
@@ -5,13 +7,16 @@ cp "$infile" "$tmp"
 cwd="$(pwd)"
 cd "$tmp" || exit 1
 
-"$cwd/main/sorbet" --silence-dev-message -a autocorrect-helpers.rb 2>&1
+if "$cwd/main/sorbet" --silence-dev-message -a autocorrect-helpers.rb 2>&1; then
+  echo "Expected to fail!"
+  exit 1
+fi
 
 echo
 echo --------------------------------------------------------------------------
 echo
 
-# Also cat the file, to make sure that 'extend' is only added once per class.
+# Also cat the file, to make that the autocorrect applied
 cat autocorrect-helpers.rb
 
 rm autocorrect-helpers.rb
